# swagger-config.yaml
openapi: 3.0.3
info:
  title: FinTrust Banking Platform API
  description: |
    Comprehensive API documentation for the FinTrust banking platform.
    
    ## Services
    - **Authentication Service** (Port 8080): User authentication, token management
    - **Consent Service** (Port 8081): Consent management, TPP registration, user profiles
    
    ## Authentication
    Most endpoints require Bearer token authentication. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```
    
    ## Base URLs
    - Authentication: `http://localhost:8080`
    - Consent: `http://localhost:8081`
    
  version: 1.0.0
  contact:
    name: FinTrust Development Team
    email: support@fintrust.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Authentication Service
  - url: http://localhost:8081
    description: Consent Service
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          description: Full name of the user
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john@example.com"
        password:
          type: string
          description: User's password (min 8 characters)
          example: "securePassword123"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "john@example.com"
        password:
          type: string
          description: User's password
          example: "securePassword123"
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (valid for 2.5 hours)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: JWT refresh token (valid for 7 days)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token to get new access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    LogoutRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token to invalidate
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    TokenValidationResult:
      type: object
      properties:
        valid:
          type: boolean
          description: Whether the token is valid
          example: true
        message:
          type: string
          description: Validation message
          example: "Token is valid"
        claims:
          type: object
          description: JWT claims (if token is valid)
          example:
            sub: "john@example.com"
            iat: 1234567890
            exp: 1234567890
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user ID
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john@example.com"
        role:
          type: string
          enum: [USER, ADMIN, TPP]
          description: User's role in the system
          example: "USER"
        authProvider:
          type: string
          enum: [LOCAL, GOOGLE]
          description: Authentication provider
          example: "LOCAL"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-01T00:00:00Z"

    # Consent DTOs
...
    TppRegistrationRequest:
      type: object
      required:
        - tppId
        - name
        - jurisdiction
        - kycDocs
        - requestedScopes
      properties:
        tppId:
          type: string
          description: Unique TPP identifier
          example: "TPP001"
        name:
          type: string
          description: TPP organization name
          example: "FinTech Solutions Ltd"
        jurisdiction:
          type: string
          description: Legal jurisdiction
          example: "India"
        kycDocs:
          type: array
          items:
            type: string
          description: IPFS hashes of KYC documents
          example: ["ipfs://QmHash1", "ipfs://QmHash2"]
        requestedScopes:
          type: array
          items:
            type: string
          description: Requested API scopes
          example: ["account_info", "transactions"]
    TppApprovalRequest:
      type: object
      required:
        - tppId
      properties:
        tppId:
          type: string
          description: TPP ID to approve
          example: "TPP001"
        adminId:
          type: string
          description: Admin ID (auto-filled)
          example: "admin@123"
        approved:
          type: boolean
          description: Approval status
          example: true
        expiryDays:
          type: integer
          description: Days until approval expires
          example: 365
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
          example: true
        message:
          type: string
          description: Response message
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Invalid input data"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-01T00:00:00Z"
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: string
                example: "User registered successfully"
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive access/refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: string
                example: "Invalid credentials"
        '500':
          description: Login failed
          content:
            application/json:
              schema:
                type: string
                example: "Login failed: Internal server error"
  /api/auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New tokens generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/auth/validate-token:
    post:
      tags:
        - Authentication
      summary: Validate JWT token
      description: Validate the provided JWT access token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenValidationResult'
        '400':
          description: Invalid Authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenValidationResult'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout user and invalidate refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/tpp/register:
    post:
      tags:
        - TPP Management
      summary: Register TPP
      description: Register a new Third-Party Provider (TPP)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TppRegistrationRequest'
      responses:
        '200':
          description: TPP registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tpp/{tppId}/status:
    get:
      tags:
        - TPP Management
      summary: Get TPP status
      description: Get registration status of a TPP
      parameters:
        - name: tppId
          in: path
          required: true
          schema:
            type: string
          description: TPP ID
          example: "TPP001"
      responses:
        '200':
          description: TPP status
          content:
            application/json:
              schema:
                type: object
                properties:
                  tppId:
                    type: string
                    example: "TPP001"
                  name:
                    type: string
                    example: "FinTech Solutions Ltd"
                  status:
                    type: string
                    enum: [PENDING, APPROVED, REJECTED]
                    example: "APPROVED"
                  approvedAt:
                    type: string
                    format: date-time
                    example: "2024-01-01T00:00:00Z"
                  expiresAt:
                    type: string
                    format: date-time
                    example: "2025-01-01T00:00:00Z"
        '404':
          description: TPP not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/approve-tpp:
    post:
      tags:
        - Admin Management
      summary: Approve TPP
      description: Approve a TPP registration (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TppApprovalRequest'
      responses:
        '200':
          description: TPP approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/pending-tpps:
    get:
      tags:
        - Admin Management
      summary: Get pending TPPs
      description: Get all pending TPP registrations (Admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of pending TPPs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    tppId:
                      type: string
                      example: "TPP001"
                    name:
                      type: string
                      example: "FinTech Solutions Ltd"
                    jurisdiction:
                      type: string
                      example: "India"
                    requestedScopes:
                      type: array
                      items:
                        type: string
                      example: ["account_info", "transactions"]
                    submittedAt:
                      type: string
                      format: date-time
                      example: "2024-01-01T00:00:00Z"
...